refactor: change structure of the project and add Webpack
