+++ src/css/style.css
@@ -7,7 +7,7 @@ body {
   background: radial-gradient(circle, #283048 0%, #859398 100%);
   background-repeat: no-repeat;
   background-size: 100% 100%;
-  height: 100%;
+  height: 100vh;
 }
 
 h1 {
