+++ src/index.html
@@ -5,10 +5,8 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Virtual keyboard</title>
-  <link rel="stylesheet" href="./css/normalize.css">
-  <link rel="stylesheet" href="./css/style.css">
 </head>
 <body>
-  <script src="./js/script.js" type="module"></script>
+
 </body>
 </html>
