+++ src/sass/_global.scss
@@ -6,7 +6,7 @@ body {
   background: radial-gradient(circle, rgba(40,48,72,1) 0%, rgba(133,147,152,1) 100%);
   background-repeat: no-repeat;
   background-size: 100% 100%;
-  height: 100%;
+  height: 100vh;
 }
 
 h1 {
