+++ src/sass/style.scss
@@ -1,3 +1,4 @@
+@import './normalize';
 @import './global';
 @import './keyboard';
 @import './key';
